feat: refactor CoursesPage to redirect to courses list and remove course management features
